# Make path for reading many parquet files without prescan