For fastparquet when there is no `_metadata`.

Fixes #3974

Benchmarks for a small dataset of 100 files, 6col*10row each, all int

Master
```
In [8]: %timeit d2 = dd.read_parquet('out.parq/')
13.3 ms ± 63.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [9]: %timeit d2 = dd.read_parquet('out.parq/*.parquet')
43.5 ms ± 135 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

this branch
```
In [4]: %timeit d2 = dd.read_parquet('out.parq/')
13.5 ms ± 251 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [5]: %timeit d2 = dd.read_parquet('out.parq/*.parquet')
4.56 ms ± 39.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
Note that reading the metadata is now *longer*, because is has to parse all of the row-groups up front (even though they are loaded from one place), whereas this is deferred to the lazy/parallel task when supplying a glob/list of paths.

This code is rather long and convoluted.