@martindurant 2018-09-13
Only failure is dask\tests\test_multiprocessing.py::test_pickle_globals , a load of extra global keys turns up.