@martindurant 2018-09-18
Any thoughts here? I think this is a thing worth doing, and, in addition, gives users the choice to supply a glob even when there is a `_metadata` file, for possibly faster response and smaller size of the serialised ParquetFile.